CRUD Bairro

+++ src/main/java/br/com/api/repository/BairroRepository.java
@@ -3,5 +3,5 @@ package br.com.api.repository;
 import br.com.api.model.Bairro;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BairroRepository extends JpaRepository<Bairro, Long> {
+public interface BairroRepository extends JpaRepository<Bairro, Integer> {
 }
